fix: Fixes grammar and implementation test mismatch
While implementing a message templates parser I pulled the message
templates tests from the C# repo to ensure that my parsing behaviour was
correct & translated them into rust. While I was doing that I saw that
alignment was specified to be able to be any digits preceeded by an
optional '-' but the test "ZeroValuesAlignmentIsParsedAsText". Since
this seems to be the most common behaviour across the ecosystem and
there's a specific test for it I expected that this was an oversight in
the spec.